requirements.txt
anthropic>=0.42.0
pandas>=2.0.0
numpy>=1.24.0
mplfinance>=0.12.10b0
matplotlib>=3.7.0
yfinance>=0.2.30
Pillow>=10.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
